fix: cast lm_head logits to FP32 before argmax + gate rep detection
Root cause: BF16 logits from lm_head have only 7 mantissa bits. When
two logit values are close, cuBLAS BF16 GEMM non-deterministic rounding
can swap the argmax winner across runs. Once a wrong token enters a
loop-inducing state, repetition is self-reinforcing.

Fix: .float() on lm_head output (all 3 models) + defense-in-depth
.float() in sampling.py greedy paths. Matches HuggingFace reference
and sglang practice.

Also: gate periodic repetition detection behind BATCHGEN_REP_DETECTION=1
(default off) — the FP32 fix is the proper prevention.

+++ batchgen/models/minimax/minimax_m25/model.py
@@ -1694,7 +1694,7 @@ class MiniMaxM25(nn.Module):
         )
 
         hidden_states = outputs[0]
-        logits = self.lm_head(hidden_states)
+        logits = self.lm_head(hidden_states).float()
 
         return CausalLMOutputWithPast(
             loss=None,
